+++ backend/clouddm-plugins/clouddm-ds/ds-mongodb/src/main/java/com/clougence/clouddm/ds/mongodb/dsconf/MongoConfig.java
@@ -21,6 +21,7 @@ import com.clougence.clouddm.base.metadata.ds.ConfigDef;
 import com.clougence.clouddm.base.metadata.ds.DataSourceConfig;
 import com.clougence.clouddm.base.metadata.ds.DataSourceType;
 import com.clougence.clouddm.base.metadata.ds.DsConfigGroup;
+import com.clougence.clouddm.base.metadata.ds.SslMode;
 import com.clougence.clouddm.ds.mongodb.execute.jdbc.MongoKeys;
 import com.clougence.clouddm.ds.mongodb.i18n.MongoConfigI18nKeys;
 import com.clougence.clouddm.sdk.execute.dsconf.Serialization;
@@ -72,6 +73,19 @@ public class MongoConfig extends DataSourceConfig {
         properties.setProperty(DsConfigKeys.DEFAULT_SCHEMA.getConfigKey(), safeStr(this.getDefaultSchema()));
         properties.setProperty(DsConfigKeys.CONNECT_TIMEOUT_MS.getConfigKey(), safeStr(StringUtils.toString(this.getConnectTimeoutMs())));
         properties.setProperty(DsConfigKeys.SO_TIMEOUT_SEC.getConfigKey(), safeStr(StringUtils.toString(this.getSoTimeoutSec())));
+
+        SslMode sslMode = this.getSslMode();
+        if (sslMode == null) {
+            sslMode = SslMode.DISABLED;
+        }
+        properties.setProperty(MongoKeys.SSL_MODE, sslMode.name());
+        properties.setProperty(MongoKeys.SSL_CA_FILE, safeStr(this.getSslCaFilePath()));
+        properties.setProperty(MongoKeys.SSL_CA_FORMAT, safeStr(this.getSslCaFileFormat()));
+        properties.setProperty(MongoKeys.SSL_CA_PASSWORD, safeStr(this.getSslCaPassword()));
+        properties.setProperty(MongoKeys.SSL_CLIENT_CERT_FILE, safeStr(this.getSslClientCertFilePath()));
+        properties.setProperty(MongoKeys.SSL_CLIENT_CERT_FORMAT, safeStr(this.getSslClientCertFileFormat()));
+        properties.setProperty(MongoKeys.SSL_CLIENT_KEY_FILE, safeStr(this.getSslClientKeyFilePath()));
+        properties.setProperty(MongoKeys.SSL_CLIENT_KEY_PASSWORD, safeStr(this.getSslClientKeyPassword()));
         return properties;
     }
 }
