+++ backend/clouddm-plugins/clouddm-ds/ds-mongodb/src/main/java/com/clougence/clouddm/ds/mongodb/execute/jdbc/MongoConnectionFactory.java
@@ -17,6 +17,7 @@ package com.clougence.clouddm.ds.mongodb.execute.jdbc;
 
 import java.sql.Connection;
 import java.sql.SQLException;
+import java.util.Map;
 import java.util.Properties;
 
 import com.clougence.clouddm.ds.mongodb.dsconf.MongoConnectType;
@@ -25,6 +26,7 @@ import com.clougence.drivers.adapter.AdapterTypeSupport;
 import com.clougence.drivers.adapter.JdbcDriver;
 import com.clougence.drivers.adapter.TypeSupport;
 import com.clougence.utils.StringUtils;
+import com.clougence.utils.ref.LinkedCaseInsensitiveMap;
 import com.mongodb.ConnectionString;
 import com.mongodb.MongoClientSettings;
 import com.mongodb.client.MongoClient;
@@ -45,7 +47,9 @@ public class MongoConnectionFactory implements AdapterFactory {
     public String[] getPropertyNames() {
         return new String[] { MongoKeys.SERVER, MongoKeys.ADAPTER_NAME, MongoKeys.DRIVER_VERSION, MongoKeys.INTERCEPTOR, MongoKeys.TIME_ZONE, MongoKeys.CONN_TIMEOUT,
                               MongoKeys.SO_TIMEOUT, MongoKeys.USERNAME, MongoKeys.PASSWORD, MongoKeys.DATABASE, MongoKeys.CLIENT_NAME, MongoKeys.MAX_TOTAL, MongoKeys.MAX_IDLE,
-                              MongoKeys.MIN_IDLE, MongoKeys.TEST_WHILE_IDLE, MongoKeys.CONNECT_TYPE };
+                              MongoKeys.MIN_IDLE, MongoKeys.TEST_WHILE_IDLE, MongoKeys.CONNECT_TYPE, MongoKeys.SSL_MODE, MongoKeys.SSL_CA_FILE, MongoKeys.SSL_CA_FORMAT,
+                              MongoKeys.SSL_CA_PASSWORD, MongoKeys.SSL_CLIENT_CERT_FILE, MongoKeys.SSL_CLIENT_CERT_FORMAT, MongoKeys.SSL_CLIENT_KEY_FILE,
+                              MongoKeys.SSL_CLIENT_KEY_PASSWORD };
     }
 
     @Override
@@ -84,6 +88,10 @@ public class MongoConnectionFactory implements AdapterFactory {
         mongoUrl = mongoUrl + optionSeparator + "connectTimeoutMS=" + connTimeoutMs + "&socketTimeoutMS=" + soTimeoutMs;
 
         MongoClientSettings.Builder settingsBuilder = MongoClientSettings.builder().applicationName(clientName).applyConnectionString(new ConnectionString(mongoUrl));
+
+        Map<String, String> caseProps = new LinkedCaseInsensitiveMap<>();
+        props.forEach((k, v) -> caseProps.put((String) k, v == null ? "" : String.valueOf(v)));
+        MongoSslFactory.apply(settingsBuilder, caseProps);
 
         MongoClient client = MongoClients.create(settingsBuilder.build());
         return new MongoConnection(owner, client, jdbcUrl, props, props.getProperty(MongoKeys.DATABASE));
